use a bigger integer type for order of events to allow AAAAMMDDHHMM as index value (#8)

+++ src-prog-maj/uMain.pas
@@ -120,8 +120,8 @@ begin
   Planning := tplanning.CreateFromURL(CPlanningServerURL, APIReadyEvent,
     APIErrorEvent);
 
-  edtOrder.Min := low(integer);
-  edtOrder.Max := high(integer);
+  edtOrder.Min := low(int64);
+  edtOrder.Max := high(int64);
 end;
 
 procedure TfrmMain.FormDestroy(Sender: TObject);
